# specs/001-ratemyprof-india-platform/contracts/api-specification.yaml
openapi: 3.0.3
info:
  title: RateMyProf India API
  description: API for rating and discovering professors in Indian colleges
  version: 0.1.0
  contact:
    name: RateMyProf India
    url: https://github.com/NihaallX/ratemyprof.in
servers:
  - url: https://api.ratemyprof.in/v1
    description: Production server
  - url: http://localhost:8000/v1
    description: Development server
paths:
...
  /reviews:
    post:
      summary: Submit a review for a professor
      operationId: createReview
      tags: [reviews]
      security:
        - bearerAuth: []
        - {}  # Allow anonymous reviews
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewCreate'
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '409':
          description: User already reviewed this professor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /reviews/{review_id}:
    put:
      summary: Update an existing review
      operationId: updateReview
      tags: [reviews]
      security:
        - bearerAuth: []
      parameters:
        - name: review_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewUpdate'
      responses:
        '200':
          description: Review updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /reviews/{review_id}/flag:
    post:
      summary: Flag a review for moderation
      operationId: flagReview
      tags: [reviews]
      security:
        - bearerAuth: []
      parameters:
        - name: review_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewFlag'
      responses:
        '201':
          description: Review flagged successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "Review flagged for moderation"
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: User already flagged this review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Authentication
...
  /moderation/reviews:
    get:
      summary: Get flagged reviews for moderation
      operationId: getFlaggedReviews
      tags: [moderation]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          description: Filter by review status
          schema:
            type: string
            enum: [pending, approved, removed]
            default: pending
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Flagged reviews for moderation
          content:
            application/json:
              schema:
                type: object
                properties:
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/FlaggedReview'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    College:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          maxLength: 200
        city:
          type: string
          maxLength: 100
        state:
          type: string
          maxLength: 50
        college_type:
          type: string
          enum: [university, college, institute]
        established_year:
          type: integer
          minimum: 1800
          maximum: 2025
          nullable: true
      required: [id, name, city, state, college_type]
    ProfessorSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
          maxLength: 50
        last_name:
          type: string
          maxLength: 50
        known_as:
          type: string
          maxLength: 50
          nullable: true
        department:
          type: string
          maxLength: 100
        college:
          $ref: '#/components/schemas/College'
        avg_ratings:
          $ref: '#/components/schemas/AverageRatings'
        total_reviews:
          type: integer
          minimum: 0
      required: [id, first_name, last_name, department, college, avg_ratings, total_reviews]
...
    AverageRatings:
      type: object
      properties:
        clarity:
          type: number
          format: float
          minimum: 1.0
          maximum: 5.0
          nullable: true
        helpfulness:
          type: number
          format: float
          minimum: 1.0
          maximum: 5.0
          nullable: true
        workload:
          type: number
          format: float
          minimum: 1.0
          maximum: 5.0
          nullable: true
        engagement:
          type: number
          format: float
          minimum: 1.0
          maximum: 5.0
          nullable: true
        overall:
          type: number
          format: float
          minimum: 1.0
          maximum: 5.0
          nullable: true
    Review:
      type: object
      properties:
        id:
          type: string
          format: uuid
        professor_id:
          type: string
          format: uuid
        author_name:
          type: string
          maxLength: 100
          description: "Display name or 'Anonymous'"
        ratings:
          type: object
          properties:
            clarity:
              type: integer
              minimum: 1
              maximum: 5
            helpfulness:
              type: integer
              minimum: 1
              maximum: 5
            workload:
              type: integer
              minimum: 1
              maximum: 5
            engagement:
              type: integer
              minimum: 1
              maximum: 5
          required: [clarity, helpfulness, workload, engagement]
        review_text:
          type: string
          maxLength: 2000
          nullable: true
        semester_taken:
          type: string
          maxLength: 50
          nullable: true
        course_taken:
          type: string
          maxLength: 100
          nullable: true
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        is_flagged:
          type: boolean
        flags_count:
          type: integer
          minimum: 0
      required: [id, professor_id, author_name, ratings, created_at, updated_at, is_flagged, flags_count]

    # Request Models
    ReviewCreate:
      type: object
      properties:
        professor_id:
          type: string
          format: uuid
        ratings:
          type: object
          properties:
            clarity:
              type: integer
              minimum: 1
              maximum: 5
            helpfulness:
              type: integer
              minimum: 1
              maximum: 5
            workload:
              type: integer
              minimum: 1
              maximum: 5
            engagement:
              type: integer
              minimum: 1
              maximum: 5
          required: [clarity, helpfulness, workload, engagement]
        review_text:
          type: string
          maxLength: 2000
          nullable: true
        semester_taken:
          type: string
          maxLength: 50
          nullable: true
        course_taken:
          type: string
          maxLength: 100
          nullable: true
        anonymous:
          type: boolean
          default: false
          description: "Whether to post anonymously"
        anon_display_name:
          type: string
          maxLength: 50
          nullable: true
          description: "Custom anonymous name (e.g., 'CS Senior')"
      required: [professor_id, ratings]
    ReviewUpdate:
      type: object
      properties:
        ratings:
          type: object
          properties:
            clarity:
              type: integer
              minimum: 1
              maximum: 5
            helpfulness:
              type: integer
              minimum: 1
              maximum: 5
            workload:
              type: integer
              minimum: 1
              maximum: 5
            engagement:
              type: integer
              minimum: 1
              maximum: 5
          required: [clarity, helpfulness, workload, engagement]
        review_text:
          type: string
          maxLength: 2000
          nullable: true
        semester_taken:
          type: string
          maxLength: 50
          nullable: true
        course_taken:
          type: string
          maxLength: 100
          nullable: true
      required: [ratings]
    ReviewFlag:
      type: object
      properties:
        reason:
          type: string
          enum: [inappropriate, spam, offensive, irrelevant, personal_attack]
        description:
          type: string
          maxLength: 500
          nullable: true
      required: [reason]
...
    FlaggedReview:
      allOf:
        - $ref: '#/components/schemas/Review'
        - type: object
          properties:
            flags:
              type: array
              items:
                type: object
                properties:
                  reason:
                    type: string
                    enum: [inappropriate, spam, offensive, irrelevant, personal_attack]
                  description:
                    type: string
                    nullable: true
                  created_at:
                    type: string
                    format: date-time
                required: [reason, created_at]
            professor:
              $ref: '#/components/schemas/ProfessorSummary'
...
    Error:
      type: object
      properties:
        error:
          type: string
          description: "Error type"
        message:
          type: string
          description: "Human-readable error message"
        details:
          type: object
          nullable: true
          description: "Additional error context"
      required: [error, message]
  responses:
    BadRequest:
      description: Invalid request data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "validation_error"
            message: "Invalid input data"
            details:
              field_errors:
                email: ["Invalid email format"]
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "authentication_required"
            message: "Valid authentication token required"
    Forbidden:
      description: Insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "insufficient_permissions"
            message: "Moderator access required"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "not_found"
            message: "Professor not found"
...